# Detailed/03-OpenAPI/med-rms-openapi.yaml
...
openapi: 3.0.3
info:
  title: Med-RMS 医疗器械需求管理系统 API
  description: 'Med-RMS (Medical Device Requirement Management System) RESTful API 规范。

    涵盖需求管理、追溯管理、变更管理、合规管理、电子签名、风险管理、项目管理、报表仪表盘及系统管理模块。

    '
  version: 1.1.0
  contact:
    name: Med-RMS 开发团队
  license:
    name: Proprietary
servers:
- url: http://localhost:8080/api/v1
  description: 开发环境
...
security:
- BearerAuth: []
paths:
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags:
      - 认证与用户
      summary: 获取当前用户信息
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/UserVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-required-permission: sys:user:list
    put:
      operationId: updateProfile
      tags:
      - 认证与用户
      summary: 更新个人信息
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileDTO'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/UserVO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /users/{id}:
    put:
      operationId: updateUser
      tags:
      - 用户管理
      summary: 更新用户
      parameters:
      - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateDTO'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/UserVO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-required-permission: sys:user:list
...
  /requirements/import:
    post:
      operationId: importRequirements
      tags:
      - 需求管理
      summary: 批量导入需求
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Excel/CSV文件
                projectId:
                  type: string
                  description: 目标项目ID
      responses:
        '200':
          description: 导入完成
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/ImportResultVO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-required-permission: req:create
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserIdParam:
      name: id
      in: path
      required: true
      description: 用户ID
      schema:
        type: string
        format: int64
...
  responses:
    BadRequest:
      description: 请求参数校验失败
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未认证（Token缺失或过期）
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 无权限访问
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 资源冲突
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: 系统内部错误
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          description: 响应状态码
          example: 200
        message:
          type: string
          description: 响应消息
          example: success
        timestamp:
          type: string
          format: date-time
          description: 响应时间戳
...
    ErrorResponse:
      type: object
      required:
      - errorCode
      - errorMessage
      properties:
        errorCode:
          type: string
          description: 错误码
          example: SY0100
        errorMessage:
          type: string
          description: 错误消息
          example: 请求参数校验失败
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          description: 字段级错误详情
    ImportResultVO:
      type: object
      required:
      - total
      - success
      - failed
      - errors
      properties:
        total:
          type: integer
          description: 总记录数
        success:
          type: integer
          description: 成功导入数
        failed:
          type: integer
          description: 失败数
        errors:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
                description: 行号
              message:
                type: string
                description: 错误消息
          description: 错误详情列表
...
    ProfileDTO:
      type: object
      properties:
        realName:
          type: string
          description: 真实姓名
          maxLength: 50
        email:
          type: string
          description: 邮箱
          format: email
          maxLength: 100
        phone:
          type: string
          description: 手机号
          maxLength: 20
        avatar:
          type: string
          description: 头像URL
...
    UserUpdateDTO:
      type: object
      properties:
        realName:
          type: string
          description: 真实姓名
          maxLength: 50
        email:
          type: string
          description: 邮箱
          format: email
        phone:
          type: string
          description: 手机号
          maxLength: 20
        orgId:
          type: string
          format: int64
          description: 组织ID
        roleIds:
          type: array
          items:
            type: string
            format: int64
          description: 角色ID列表
    UserVO:
      type: object
      properties:
        id:
          type: string
          format: int64
          description: 用户ID
        username:
          type: string
          description: 用户名
        realName:
          type: string
          description: 真实姓名
        email:
          type: string
          description: 邮箱
        phone:
          type: string
          description: 手机号
        avatar:
          type: string
          description: 头像URL
        status:
          type: string
          enum:
          - Active
          - Disabled
          description: 状态
        orgId:
          type: string
          format: int64
          description: 组织ID
        orgName:
          type: string
          description: 组织名称
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleVO'
          description: 角色列表
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    RoleVO:
      type: object
      properties:
        id:
          type: string
          format: int64
          description: 角色ID
        name:
          type: string
          description: 角色名称
        code:
          type: string
          description: 角色编码
        description:
          type: string
          description: 角色描述
        permissions:
          type: array
          items:
            type: string
          description: 权限代码列表
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...